requirements.txt
Flask
Flask-Session
requests
python-dotenv
gunicorn
psycopg2-binary
discord.py
aiohttp
google-genai
matplotlib
pillow
pytz
flask-socketio
eventlet
